+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 1.0.3
+- Updated KHQR Card Widget
+- Added thousand number formatter util
+- Allow to keep decimal places
+- Updated dependencies
+
 ## 1.0.2
 - Updated swift package manager support
 - Updated KHQR Card Widget